requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
pyautogen>=0.2.0
python-chess>=1.999
typing-extensions>=4.8.0
pydantic>=2.0.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
python-dotenv>=1.0.0

